# swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a server, which proides api for signup, signin user and get information about user.
    Developers LastUnion team.
  version: 1.0.0
  title: LastUnion server
# schemes:
# - https

paths:
  /api/user/signin:
    post:
      tags:
      - uesr
      summary: Sign in user into the server
      description: Sign in user into the server
      operationId: signInUser
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/signInUserModel'
      responses:
        200:
          description: Successfull operation
          schema:
            $ref: '#/definitions/Ok'
        400:
          description: Contains invalid fields
          schema:
            $ref: '#/definitions/badRequest'
        403:
          description: Invalid login or password
          schema: 
            $ref: '#/definitions/forbidden'
  /api/user/signup:
    post:
      tags:
      - user
      summary: Create user 
      operationId: signUpUser
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/signUpUserModel'
      responses:
        201:
          description: Successfull signup
          schema:
            $ref: '#/definitions/created'
        400:
          description: Contains invalid fields
          schema:
            $ref: '#/definitions/badRequest'
        409:
          description: Login or Email already registered
          schema: 
            $ref: '#/definitions/conflict'
  /api/user/data:
    get:
      tags:
      - user
      summary: Get information about user
      operationId: getUserData
      produces: 
      - application/json
      parameters: []
      responses:
        200:
          description: Successfull operation
          schema:
             $ref: '#/definitions/Ok'
        403:
          description: Invalid login or password
          schema: 
            $ref: '#/definitions/forbidden'
        404:
          description: Invalid session
          schema: 
            $ref: '#/definitions/notFound'
  /api/user/change_email:
    post:
      tags:
      - user
      summary: Changes user email
      operationId: changeUserEmail
      consumes:
      - application/json
      produces: 
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/changeUserEmailModel'
      responses:
        200:
          description: Successfull operation
          schema:
             $ref: '#/definitions/Ok'
        400:
          description: Contains invalid fields
          schema:
            $ref: '#/definitions/badRequest'
        404:
          description: Invalid sesion
          schema: 
            $ref: '#/definitions/notFound'
        409:
          description: Login or Email already registered
          schema: 
            $ref: '#/definitions/conflict'
            
  /api/user/change_password:
    post:
      tags:
      - user
      summary: Changes user password
      operationId: changeUserPassword
      consumes:
      - application/json
      produces: 
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/changeUserPasswordModel'
      responses:
        200:
          description: Successfull operation
          schema:
             $ref: '#/definitions/Ok'
        400:
          description: Contains invalid fields
          schema:
            $ref: '#/definitions/badRequest'
        404:
          description: Invalid session
          schema: 
            $ref: '#/definitions/notFound'
        403:
          description: Invalid login or password
          schema: 
            $ref: '#/definitions/forbidden'
  /api/user/delete:
    delete:
      tags:
      - user
      summary: delete user
      operationId: deleteUser
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: Successfull operation
          schema:
            $ref: '#/definitions/Ok'
        404:
          description: Invalid sesion
          schema: 
            $ref: '#/definitions/notFound'
  /api/user/logout:
    post:
      tags:
      - user
      summary: logout user
      operationId: logoutUser
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: Successfull operation
          schema:
             $ref: '#/definitions/Ok'
        404:
          description: Invalid session
          schema: 
            $ref: '#/definitions/notFound'
      
definitions:
  signInUserModel:
    type: object
    properties:
      userName:
        type: string
        format: unique
      userPassword:
        type: string
        format: password
  signUpUserModel:
    type: object
    properties:
      userName:
        type: string
        format: unique
      userEmail:
        type: string
        format: unique
      userPassword:
        type: string
        format: password
  changeUserEmailModel:
    type: object
    properties:
      userEmail:
        type: string
        format: unique
  changeUserPasswordModel:
    type: object
    properties:
      oldPassword:
        type: string
        format: password
      newPassword:
        type: string
        format: password
  Ok:
    type: object
    properties:
      result: 
        type: boolean
        example: true
      responseMessage:
        type: string
        example: Ok! en
      data: 
        type: string
        example: null
    xml: 
      name: Ok
  created:
    type: object
    properties:
      result: 
        type: boolean
        example: true
      responseMessage:
        type: string
        example: User created successfully! en
      data: 
        type: string
        example: null
  badRequest:
    type: object
    properties:
      result: 
        type: boolean
        example: false
      responseMessage:
        type: string
        example: Json contains null fields! en
      data: 
        type: string
        example: null
  forbidden:
    type: object
    properties:
      result: 
        type: boolean
        example: false
      responseMessage:
        type: string
        example: Invalid authentication data! en
      data: 
        type: string
        example: null
  notFound:
    type: object
    properties:
      result: 
        type: boolean
        example: false
      responseMessage:
        type: string
        example: Invalid session! en
      data: 
        type: string
        example: null
  conflict:
    type: object
    properties:
      result: 
        type: boolean
        example: false
      responseMessage:
        type: string
        example: Login already occupied! en
      data: 
        type: string
        example: null
    
      
    
    
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
# basePath: /HustonMmmavr/server_api/1.0.0
schemes:
 - https
 - http
# Added by API Auto Mocking Plugin
basePath: /HustonMmmavr/new/1.0.0